requirements.txt
# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Spark
pyspark==3.5.0

# Database clients
psycopg2-binary==2.9.9
elasticsearch==8.11.0
redis==5.0.1

# ML & Data Science
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Orchestration
apache-airflow==2.7.3

# Monitoring & Logging
mlflow==2.8.1
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0

# Development
pytest==7.4.3
jupyter==1.0.0
black==23.11.0
